requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
addict=2.4.0=pypi_0
aiohappyeyeballs=2.4.6=pypi_0
aiohttp=3.11.13=pypi_0
aiosignal=1.3.2=pypi_0
albucore=0.0.23=pypi_0
albumentations=2.0.4=pypi_0
altair=5.5.0=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
attrs=25.1.0=pypi_0
black=25.1.0=pypi_0
blinker=1.9.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
cachetools=5.5.2=pypi_0
certifi=2025.1.31=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
clip=1.0=pypi_0
diffusers=0.32.2=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.8.1=pypi_0
filelock=3.17.0=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2025.2.0=pypi_0
ftfy=6.3.1=pypi_0
future=1.0.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
grpcio=1.70.0=pypi_0
huggingface-hub=0.29.1=pypi_0
idna=3.10=pypi_0
imageio=2.37.0=pypi_0
imageio-ffmpeg=0.6.0=pypi_0
importlib-metadata=8.6.1=pypi_0
invisible-watermark=0.2.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.5=pypi_0
joblib=1.4.2=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
kornia=0.8.0=pypi_0
kornia-rs=0.1.8=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
lightning=2.5.0.post0=pypi_0
lightning-utilities=0.12.0=pypi_0
lpips=0.1.4=pypi_0
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
mypy-extensions=1.0.0=pypi_0
narwhals=1.28.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.4.2=pypi_0
numpy=2.2.3=pypi_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
omegaconf=2.3.0=pypi_0
opencv-python=4.11.0.86=pypi_0
opencv-python-headless=4.11.0.86=pypi_0
openssl=3.0.15=h5eee18b_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
parso=0.8.4=pypi_0
pathspec=0.12.1=pypi_0
pillow=11.1.0=pypi_0
pip=25.0=py311h06a4308_0
platformdirs=4.3.6=pypi_0
propcache=0.3.0=pypi_0
protobuf=5.29.3=pypi_0
psutil=7.0.0=pypi_0
pudb=2024.1.3=pypi_0
pyarrow=19.0.1=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.19.1=pypi_0
python=3.11.11=he870216_0
python-dateutil=2.9.0.post0=pypi_0
pytorch-lightning=2.5.0.post0=pypi_0
pytz=2025.1=pypi_0
pywavelets=1.8.0=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rich=13.9.4=pypi_0
rpds-py=0.23.1=pypi_0
safetensors=0.5.3=pypi_0
scikit-image=0.25.2=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.2=pypi_0
sentry-sdk=2.22.0=pypi_0
setproctitle=1.3.5=pypi_0
setuptools=75.8.0=py311h06a4308_0
simsimd=6.2.1=pypi_0
six=1.17.0=pypi_0
smmap=5.0.2=pypi_0
sqlite=3.45.3=h5eee18b_0
streamlit=1.42.2=pypi_0
stringzilla=3.12.1=pypi_0
sympy=1.13.1=pypi_0
taming-transformers=0.0.1=pypi_0
tenacity=9.0.0=pypi_0
tensorboard=2.19.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboardx=2.6.2.2=pypi_0
test-tube=0.7.5=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2025.2.18=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.21.0=pypi_0
toml=0.10.2=pypi_0
tomli=2.2.1=pypi_0
torch=2.5.0+cu124=pypi_0
torch-fidelity=0.3.0=pypi_0
torchmetrics=1.6.1=pypi_0
torchvision=0.20.0+cu124=pypi_0
tornado=6.4.2=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.49.0=pypi_0
triton=3.1.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025.1=pypi_0
urllib3=2.3.0=pypi_0
urwid=2.6.16=pypi_0
urwid-readline=0.15.1=pypi_0
wandb=0.19.7=pypi_0
watchdog=6.0.0=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py311h06a4308_0
xz=5.6.4=h5eee18b_1
yapf=0.40.1=pypi_0
yarl=1.18.3=pypi_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h5eee18b_1
